Fill method test, 2019-05-22 7:58PM
In command line, compile and run PictureExplorer.java. Clicking on a pixel should fill in any connected pixel that is the same colour. The fill method is in the Picture class.

The fill method uses a queue to store Pixels to operate on later. Recursion cannot be used due to limited stack memory.

PROBLEM: the fill method doesn't work properly. In the IF statement choosing the neighbouring pixels to check, the program acts as though the last ELSE IF statement isn't executed. (In the uploaded version, the last else if statement checks the pixel to the left (westPixel), which is never filled in. I tested this with eastPixel as well.) However, the fill method sometimes fails beyond just that.